% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigPint.R
\docType{package}
\name{bigPint}
\alias{bigPint}
\alias{bigPint-package}
\title{\code{bigPint} package}
\description{
bigPint R API
}
\details{
See the README on
\href{https://github.com/lrutter/bigPint#readme}{GitHub}
}
