% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSEPair.R
\name{convertSEPair}
\alias{convertSEPair}
\title{Convert SummarizedExperiment to contain defined treatment pair}
\usage{
convertSEPair(dataSE, group1, group2)
}
\arguments{
\item{dataSE}{SUMMARIZEDEXPERIMENT | Summarized experiment format that
can be used in lieu of data and dataMetrics}

\item{group1}{CHARACTER STRING | Name of one treatment group that will
remain in the dataSE object}

\item{group2}{CHARACTER STRING | Name of second treatment group that will
remain in the dataSE object}
}
\value{
A new dataSE object that is a subset of the input dataSE in that it
now only contains the user-specified pair of treatment groups.
}
\description{
Reduce a SummarizedExperiment object that initially contains
more than two treatment groups to now only contain a user-specified subset
pair of treatment groups. Note that this function is only necesary for
users using the SummarizedExperiment input object (instead of the
combination of data and dataMetrics input objects.)
}
\examples{
# Example: Read in example SummarizedExperiment object that contains three
# treatment groups (S1, S2, and S3). Reduce it to now only contain two
# treatment groups (S1 and S3).

data(se_soybean_cn_sub)
se_soybean_cn_sub_2 <- convertSEPair(se_soybean_cn_sub, "S1", "S3")

}
