% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSESubsetGenes.R
\name{convertSESubsetGenes}
\alias{convertSESubsetGenes}
\title{Convert SummarizedExperiment to only contain defined genes}
\usage{
convertSESubsetGenes(dataSE, geneList)
}
\arguments{
\item{dataSE}{SUMMARIZEDEXPERIMENT | Summarized experiment format that
can be used in lieu of data and dataMetrics}

\item{geneList}{CHARACTER ARRAY | List of gene IDs to remain in dataSE}
}
\value{
A new dataSE object that is a subset of the input dataSE in that it
now only contains the user-specified list of genes.
}
\description{
Reduce a SummarizedExperiment object so that it only now
contains a subset of genes.
}
\examples{
Example: Read in example SummarizedExperiment object that originally
contains 5604 genes. Reduce it to now only contain ten genes (specifically
the ones with the lowest FDR).

data(se_soybean_ir_sub)
geneList <- as.data.frame(rowData(se_soybean_ir_sub)) \%>\%
    arrange(N_P.FDR) \%>\% filter(row_number() <= 10)
geneList <- geneList[,1]
se_soybean_ir_sub_2 <- convertSESubsetGenes(se_soybean_ir_sub, geneList)

}
