% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{degFC}
\alias{degFC}
\title{Superimpose DEGs onto scatterplot matrix fold change}
\usage{
degFC(data, dataMetrics, threshFC = threshFC, threshVar = threshVar,
  threshVal = threshVal, bluePointSize = bluePointSize,
  redPointSize = redPointSize, greyPointSize = greyPointSize,
  outDir = outDir)
}
\description{
Superimpose DEGs onto scatterplot matrix fold change
}
