% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{degPCP}
\alias{degPCP}
\title{Plot DEGs as parallel coordinate plots}
\usage{
degPCP(data, dataMetrics, threshVar = threshVar, threshVal = threshVal,
  lineList = lineList, lineSize = lineSize, lineColor = lineColor,
  outDir = outDir, fileName = fileName)
}
\description{
Plot DEGs as parallel coordinate plots
}
