% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{degPI}
\alias{degPI}
\title{Superimpose DEGs onto scatterplot matrix of prediction intervals}
\usage{
degPI(data, dataMetrics, threshVar = threshVar, threshVal = threshVal,
  piLevel = piLevel, bluePointSize = bluePointSize,
  redPointSize = redPointSize, greyPointSize = greyPointSize,
  outDir = outDir)
}
\description{
Superimpose DEGs onto scatterplot matrix of prediction intervals
}
