% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{degScatMat}
\alias{degScatMat}
\title{Superimpose DEGs onto scatterplot matrix}
\usage{
degScatMat(data = data, dataMetrics = dataMetrics, pointSize = pointSize,
  xbins = xbins, threshVar = threshVar, threshVal = threshVal,
  outDir = outDir)
}
\description{
Superimpose DEGs onto scatterplot matrix
}
