% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{degScatMatPoints}
\alias{degScatMatPoints}
\title{Superimpose DEGs onto scatterplot matrix}
\usage{
degScatMatPoints(data = data, dataMetrics = dataMetrics,
  pointSize = pointSize, degPointColor = degPointColor,
  threshVar = threshVar, threshVal = threshVal, outDir = outDir,
  degScatMatPoints = degScatMatPoints, fileName = fileName)
}
\description{
Superimpose DEGs onto scatterplot matrix
}
