% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{degVolcano}
\alias{degVolcano}
\title{Superimpose DEGs onto volcano plot}
\usage{
degVolcano(data, dataMetrics, logFC = logFC, PValue = PValue,
  threshVar = threshVar, threshVal = threshVal, xbins = xbins,
  pointSize = pointSize, outDir = outDir)
}
\description{
Superimpose DEGs onto volcano plot
}
