% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusters.R
\name{plotClusters}
\alias{plotClusters}
\title{Plot parallel coordinate lines for clusters}
\usage{
plotClusters(data, dataMetrics, nC, threshVar = "FDR", threshVal = 0.05,
  topNum = -1, outDir = getwd(), verbose = FALSE)
}
\arguments{
\item{data}{data frame containing read counts}

\item{dataMetrics}{data frame containing metrics}

\item{nC}{the number of clusters}

\item{threshVar}{the name of column in dataMetrics object that is used to threshold significance (character string; default "FDR")}

\item{threshVal}{the maximum value of threshVar from which to select genes to cluster (default 0.05)}

\item{outDir}{output directory to save all images (default current directory)}

\item{verbose}{in addition to the usual collective printing of all clusters from a given cluster size, print each cluster from each cluster size into separate images and print the associated IDs of each cluster from each cluster size into separate text files (default is FALSE)}

\item{threshNum}{the number of genes with the lowest threshVar values to select genes to cluster (default is for threshNum to equal -1 and to select clustering genes based on threshVal. If threshNum is changed to a positive value, then threshVal is overridden)}
}
\description{
Plot parallel coordinate lines for clusters
}
\examples{
data(soybean_cn)
data(soybean_cn_metrics)
for (nC in c(3,6)){plotClusters(data=soybean_cn, dataMetrics = soybean_cn_metrics, nC=nC)}
}
