% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{plotDEG}
\alias{plotDEG}
\title{Plot differentially expressed genes}
\usage{
plotDEG(data = data, dataMetrics = dataMetrics, outDir = getwd(),
  pointSize = 0.5, bluePointSize = 0.1, redPointSize = 0.1,
  greyPointSize = 0.1, lineSize = 0.1, lineColor = "orange",
  degPointColor = "orange", xbins = 10, piLevel = 0.95, threshFC = 3,
  threshOrth = 3, threshVar = "FDR", threshVal = 0.05, lineList = NULL,
  logFC = "logFC", PValue = "PValue", option = "scatterPoints",
  fileName = "")
}
\arguments{
\item{data}{data frame containing read counts}

\item{dataMetrics}{data frame containing metrics}

\item{outDir}{output directory to save all images (default current directory)}

\item{pointSize}{size of plotted points (default 0.5; used in "scatterHexagon", "scatterPoints", and "volcano")}

\item{bluePointSize}{size of plotted blue points (default 0.1; used in "scatterFoldChange" and "scatterOrthogonal" and "scatterPrediction")}

\item{redPointSize}{size of plotted red points (default 0.1; used in "scatterFoldChange" and "scatterOrthogonal" and "scatterPrediction")}

\item{greyPointSize}{size of plotted grey points (default 0.1; used in "scatterFoldChange" and "scatterOrthogonal" and "scatterPrediction")}

\item{lineSize}{size of plotted parallel coordinate lines (default 0.1; used in "parallelCoord")}

\item{lineColor}{color of plotted parallel coordinate lines (default "orange"; used in "parallelCoord")}

\item{degPointColor}{color of DEGs plotted as points on scatterplot matrix (default "orange; used in "scatterPoints")}

\item{xbins}{the number of bins partitioning the range of the plot (default 10; used in "scatterHexagon")}

\item{piLevel}{prediction interval level (between 0 and 1; default 0.95; used in "scatterPrediction")}

\item{threshFC}{threshold of fold change (default 3; used in "scatterFoldChange")}

\item{threshOrth}{threshold of orthogonal distance (default 3; used in "scatterOrthogonal")}

\item{threshVar}{name of column in dataMetrics object that is used to threshold significance (character string; default "FDR"; used in all)}

\item{threshVal}{maximum value to threshold significance from threshVar object (default 0.05; used in all)}

\item{lineList}{list of ID values of genes to be drawn from data as parallel coordinate lines. Use this parameter if you have predetermined genes to be drawn. Otherwise, use dataMetrics, threshVar, and threshVal to create genes to be drawn (default NULL; used in "parallelCoord")}

\item{logFC}{name of column in dataMetrics object that contains log fold change values (character string; default "logFC"; used in "volcano")}

\item{PValue}{name of column in dataMetrics object that contains p-values (character string; default "PValue"; used in "volcano")}

\item{option}{the type of plot (can choose from c("parallelCoord", "scatterFoldChange", "scatterHexagon", "scatterOrthogonal", "scatterPoints", "scatterPrediction", "volcano"); default "scatterPoints")}

\item{fileName}{the name of the output file (default is based on plot option)}

\item{saveFile}{save file to outDir (default FALSE)}
}
\description{
Superimpose differentially expressed genes onto background plot containing all genes.
}
\details{
There are seven options:
\itemize{
 \item{"scatterHexagon": }{Plot DEGs onto a scatterplot matrix of hexagon binning}
 \item{"scatterPoints": }{Plot DEGs onto a scatterplot matrix of points}
 \item{"scatterOrthogonal": }{Plots DEGs onto a scatterplot matrix of orthogonal distance}
 \item{"scatterFoldChange": }{Plots DEGs onto a scatterplot matrix of fold changes}
 \item{"scatterPrediction": }{Plot DEGs onto a scatterplot matrix of prediction intervals}
 \item{"parallelCoord": }{Plots DEGs as parallel coordinate plots on top of boxplot}
 \item{"volcano": }{Plot DEGs onto a volcano plot}
}
}
\examples{
data(soybean_cn)
data(soybean_cn_metrics)
plotDEG(soybean_cn, soybean_cn_metrics)
}
