% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCP.R
\name{plotPCP}
\alias{plotPCP}
\title{Highlight parallel coordinate plot lines inside selection box}
\usage{
plotPCP(pcpDat, option = "deleteInteger")
}
\arguments{
\item{pcpDat}{the data frame that contains the parallel coordinate plot values}

\item{option}{the interactivity option ("deleteInteger", "delete", "highlight"); default ("deleteInteger")}
}
\description{
Highlight parallel coordinate plot lines inside selection box
}
\examples{
set.seed(3)
f = function(){1.3*rnorm(100)}
pcpDat = data.frame(ID = paste0("ID", 1:100), A=f(), B=f(), C=f(), D=f(), E=f(), F=f())
pcpDat$ID = as.character(pcpDat$ID)
plotPCP(pcpDat = pcpDat)
}
