% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPermutations.R
\name{plotPermutations2}
\alias{plotPermutations2}
\title{data(soybean_cn)
data <- soybean_cn
plotPermutations(data, nPerm = 10, topThresh = 30, outDir = getwd())}
\usage{
plotPermutations2(data = data, nPerm = 10, topThresh = 50,
  threshVal = 0.05, outDir = getwd())
}
\description{
data(soybean_cn)
data <- soybean_cn
plotPermutations(data, nPerm = 10, topThresh = 30, outDir = getwd())
}
