% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRepLine.R
\name{plotRepLine}
\alias{plotRepLine}
\title{Replicate line plot linked with parallel coordinate plot}
\usage{
plotRepLine(data, dataMetrics)
}
\arguments{
\item{data}{the data frame that contains the logged read counts for all samples}

\item{dataMetrics}{the data frame that contains the metrics of interest (common choices include the p-value and log fold changes)}
}
\description{
Replicate line plot linked with parallel coordinate plot
}
\examples{
data(soybean_ir)
soybean_ir[,-1] <- log(soybean_ir[,-1]+1)
data(soybean_ir_metrics)
plotRepLine(soybean_ir, soybean_ir_metrics)
}
