% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterInteractive.R
\name{plotScatterInteractive}
\alias{plotScatterInteractive}
\title{Plot interactive scatterplot matrices}
\usage{
plotScatterInteractive(data = data, outDir = getwd(), threshOrth = 3,
  xbins = 10, option = "hexagon")
}
\arguments{
\item{data}{data frame containing read counts}

\item{outDir}{output directory to save all images (default current directory)}

\item{threshOrth}{threshold of orthogonal distance (default 3; used in "orthogonal")}

\item{xbins}{the number of bins partitioning the range of the plot (default 10; used in "hexagon")}

\item{option}{the type of plot (can choose from c("hexagon", "foldChange", "orthogonal", "prediction"); default "hexagon")}
}
\description{
Plot interactive scatterplot matrices.
}
\details{
There are four options:
\itemize{
 \item{"hexagon": }{Plot interactive scatterplot matrix with hexagon binning}
 \item{"foldChange": }{Plot interactive scatterplot matrix with fold change}
 \item{"orthogonal": }{Plot interactive scatterplot matrix with orthogonal distance}
 \item{"prediction": }{Plot interactive scatterplot matrix with prediction interval}
}
}
\examples{
data(soybean_cn)
soybean_cn <- soybean_cn[,1:7]
plotScatterInteractive(soybean_cn)
}
