% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterStatic.R
\name{plotScatterStatic}
\alias{plotScatterStatic}
\title{Plot static scatterplot matrices}
\usage{
plotScatterStatic(data = data, outDir = getwd(), saveFile = TRUE,
  pointSize = 1, threshFC = 3, threshOrth = 3, piLevel = 0.95,
  xbins = 10, option = "hexagon")
}
\arguments{
\item{data}{data frame containing read counts}

\item{outDir}{output directory to save all images (default current directory)}

\item{saveFile}{save file to outDir (default TRUE)}

\item{pointSize}{size of plotted points (default 1; used in "foldChange", "orthogonal", "prediction", and "point")}

\item{threshFC}{threshold for the fold change (default 3; used in "foldChange")}

\item{threshOrth}{threshold of orthogonal distance (default 3; used in "orthogonal")}

\item{piLevel}{prediction interval level (between 0 and 1; default 0.95; used in "prediction")}

\item{xbins}{the number of bins partitioning the range of the plot (default 10; used in "hexagon")}

\item{option}{the type of plot (can choose from c("hexagon", "foldChange", "orthogonal", "prediction", "point"); default "hexagon")}
}
\description{
Plot static scatterplot matrices.
}
\details{
There are four options:
\itemize{
 \item{"hexagon": }{Plot static scatterplot matrix with hexagon binning}
 \item{"foldChange": }{Plot static scatterplot matrix with fold change}
 \item{"orthogonal": }{Plot static scatterplot matrix with orthogonal distance}
 \item{"prediction": }{Plot static scatterplot matrix with prediction interval}
 \item{"point": }{Plot static scatterplot matrix with raw points}
}
}
\examples{
data(soybean_cn)
soybean_cn <- soybean_cn
plotScatterStatic(soybean_cn)
}
