% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot static volcano plot}
\usage{
plotVolcano(data = data, dataMetrics = dataMetrics,
  threshVar = "FDR", threshVal = 0.05, geneList = NULL,
  saveFile = TRUE, outDir = getwd(), pointColor = "orange",
  pointSize = 0.5, xbins = 10, logFC = "logFC", PValue = "PValue",
  option = "hexagon", hover = FALSE)
}
\arguments{
\item{data}{DATA FRAME | Read counts}

\item{dataMetrics}{LIST | Differential expression metrics}

\item{threshVar}{CHARACTER STRING | Name of column in dataMetrics object that is used to threshold significance; default "FDR"}

\item{threshVal}{INTEGER | Maximum value to threshold significance from threshVar object; default 0.05}

\item{geneList}{CHARACTER ARRAY | List of gene IDs to be drawn onto the scatterplot matrix of all data. Use this parameter if you have predetermined subset of genes to be superimposed. Otherwise, dataMetrics, threshVar, and threshVal will be used to create genes to be superimposed onto the volcano plot; default NULL}

\item{saveFile}{BOOLEAN [TRUE | FALSE] | Save file to outDir; default TRUE}

\item{outDir}{CHARACTER STRING | Output directory to save all plots; default current directory}

\item{pointColor}{CHARACTER STRING | Color of overlaid points on scatterplot matrix; default "orange"}

\item{pointSize}{INTEGER | Size of plotted points; default 0.5}

\item{xbins}{INTEGER | Number of bins partitioning the range of the plot; default 10}

\item{logFC}{CHARACTER STRING | Name of column in dataMetrics object that contains log fold change values; default "logFC"}

\item{PValue}{CHARACTER STRING | Name of column in dataMetrics object that contains p-values; default "PValue"}

\item{option}{CHARACTER STRING ["hexagon" | "allPoints"] | The background of plot; default "hexagon"}

\item{hover}{BOOLEAN [TRUE | FALSE] | Allow to hover over points to identify IDs; default FALSE}
}
\description{
Plot static volcano plot.
}
\examples{
# Example 1: Plot volcano plot with default settings for overlaid points (FDR < 0.05).

data(soybean_ir_sub)
data(soybean_ir_sub_metrics)
ret <- plotVolcano(soybean_ir_sub, soybean_ir_sub_metrics, pointSize = 1, saveFile = FALSE)
ret[[1]]

# Example 2: Plot volcano plot and overlay points with PValue < 1e-15.

ret <- plotVolcano(soybean_ir_sub, soybean_ir_sub_metrics, pointColor = "red", pointSize = 1,
  threshVar = "PValue", threshVal = 1e-15, saveFile = FALSE)
ret[[1]]

# Example 3: Plot volcano plot and overlay points with PValue < 1e-15. This time, plot all 
points (instead of hexagons) for the background.

ret <- plotVolcano(soybean_ir_sub, soybean_ir_sub_metrics, pointColor = "red", pointSize = 1,
  threshVar = "PValue", threshVal = 1e-15, option = "allPoints", saveFile = FALSE)
ret[[1]]

# Example 4: Plot volcano plot with points in background and overlay points with
PValue < 1e-15. This time, use a value of TRUE for the hover parameter so that you can hover 
over overlaid points and determine their IDs .

ret <- plotVolcano(soybean_ir_sub, soybean_ir_sub_metrics, pointColor = "red", pointSize = 1,
  threshVar = "PValue", threshVal = 1e-15, option = "allPoints", saveFile = FALSE, 
  hover = TRUE)
ret[[1]]

}
