% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCPInteractive.R
\name{selDelIntPCP}
\alias{selDelIntPCP}
\title{Delete parallel coordinate plot lines inside integers of selection box}
\usage{
selDelIntPCP(pcpDat)
}
\arguments{
\item{pcpDat}{the data frame that contains the parallel coordinate plot values}
}
\description{
Delete parallel coordinate plot lines inside integers of selection box
}
