% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCP.R
\name{selDelPCP}
\alias{selDelPCP}
\title{Delete parallel coordinate plot lines inside selection box}
\usage{
selDelPCP(pcpDat)
}
\arguments{
\item{pcpDat}{the data frame that contains the parallel coordinate plot values}
}
\description{
Delete parallel coordinate plot lines inside selection box
}
