% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCP.R
\name{selPCP}
\alias{selPCP}
\title{Highlight parallel coordinate plot lines inside selection box}
\usage{
selPCP(pcpDat)
}
\arguments{
\item{pcpDat}{the data frame that contains the parallel coordinate plot values}
}
\description{
Highlight parallel coordinate plot lines inside selection box
}
