% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soybean_cn_sub.R
\docType{data}
\name{soybean_cn_sub}
\alias{soybean_cn_sub}
\title{Normalized and subsetted soybean cotyledon data}
\format{a \code{RData} instance, 1 row per gene}
\usage{
data(soybean_cn_sub)
}
\description{
This dataset contains normalized RNA-sequencing read counts from soybean cotyledon across three time stages of development. Early stage cotyledons were collected four days after planting and were green but closed. Middle stage cotyledons were collected while green and open, soon after the plant generated its first set of unifoliate leaves. Late stage cotyledons were collected immediately after the initiation of yellowing and shrinking. To save on size, this example dataset was generated by obtaining a random subset of 1 out of 10 genes from the original resource.
}
\details{
Normalized and subsetted soybean cotyledon data

\itemize{
\item ID gene name
\item S1.1 early stage replicate 1 normalized read counts
\item S1.2 early stage replicate 2 normalized read counts
\item S1.3 early stage replicate 3 normalized read counts
\item S2.1 middle stage replicate 1 normalized read counts
\item S2.2 middle stage replicate 2 normalized read counts
\item S2.3 middle stage replicate 3 normalized read counts
\item S3.1 late stage replicate 1 normalized read counts
\item S3.2 late stage replicate 2 normalized read counts
\item S3.3 late stage replicate 3 normalized read counts
}
}
\references{
Brown AV, Hudson KA (2015) Developmental profiling of gene expression in soybean trifoliate leaves and cotyledons. BMC Plant Biol 15:169
}
\seealso{
\code{\link{soybean_cn}} from which this dataset is subsetted
}
\keyword{datasets}
