% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soybean_cn_sub_metrics.R
\docType{data}
\name{soybean_cn_sub_metrics}
\alias{soybean_cn_sub_metrics}
\title{Normalized and subsetted soybean cotyledon metrics}
\format{a \code{RData} instance, 1 list per treatment group combination and 1 row per gene}
\usage{
data(soybean_cn_sub_metrics)
}
\description{
This data contains metrics for normalized RNA-sequencing read counts from soybean cotyledon across three time stages of development. Early stage cotyledons were collected four days after planting and were green but closed. Middle stage cotyledons were collected while green and open, soon after the plant generated its first set of unifoliate leaves. Late stage cotyledons were collected immediately after the initiation of yellowing and shrinking. The metrics include the log fold change, log counts per million, likelihood ratio, p-values, and FDR values for all genes and all pairwise combinations of treatment groups. To save on size, this example dataset was generated by obtaining a random subset of 1 out of 10 genes from the original resource.
}
\details{
Normalized and subsetted soybean cotyledon metrics

\itemize{
\item ID gene name
\item logFC log fold change
\item logCPM log counts per million
\item LR likelihood ratio
\item PValue p-value
\item FDR FDR value
}
}
\seealso{
\code{\link{soybean_cn_sub}} for information about the treatment groups
}
\keyword{datasets}
