% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soybean_ir.R
\docType{data}
\name{soybean_ir}
\alias{soybean_ir}
\title{Raw soybean leaves iron-metabolism data}
\format{
a \code{RData} instance, 1 row per gene

A data frame with 56,044 rows and 7 variables
}
\description{
This dataset contains raw RNA-sequencing read counts from a 
soybean dataset that compared leaves that were exposed to iron-rich (iron
-postive) soil conditions versus leaves that were exposed to iron-poor 
(iron-negative) soil conditions. The data was collected 120 minutes after 
iron conditions were initiated.
}
\details{
Raw soybean leaves data

\itemize{
\item ID gene name
\item N.1 iron-negative condition replicate 1 raw read counts
\item N.2 iron-negative condition replicate 2 raw read counts
\item N.3 iron-negative condition replicate 3 raw read counts
\item P.1 iron-positive condition replicate 1 raw read counts
\item P.2 iron-positive condition replicate 2 raw read counts
\item P.3 iron-positive condition replicate 3 raw read counts
}
}
\keyword{datasets}
