% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLitreApp.R
\name{plotLitreApp}
\alias{plotLitreApp}
\title{Plot interactive litre plots}
\usage{
plotLitreApp(data = data, dataMetrics = dataMetrics, geneList = NULL,
  pointColor = "orange", option = "hexagon")
}
\arguments{
\item{data}{DATA FRAME | Read counts}

\item{dataMetrics}{LIST | Differential expression metrics (required)}

\item{geneList}{CHARACTER ARRAY | List of gene IDs to be drawn onto the litre. Use this parameter if you have predetermined subset of genes to be drawn. Otherwise, all genes in the data object can be superimposed on the litre plot; default NULL}

\item{pointColor}{CHARACTER STRING | Color of overlaid points on scatterplot matrix; default "orange"}

\item{option}{CHARACTER STRING ["hexagon" | "allPoints"] | The background of plot; default "hexagon"; "allPoints" may be too slow depending on data}
}
\description{
Plot interactive litre plots.
}
\examples{
\dontrun{
# Example 1: Create an interactive litre plot for the logged data using default background of 
# hexagons.

data(soybean_ir_sub)
data(soybean_ir_sub_metrics)
soybean_ir_sub_log <- soybean_ir_sub
soybean_ir_sub_log[,-1] <- log(soybean_ir_sub[,-1]+1)
plotLitreApp(data = soybean_ir_sub_log, dataMetrics = soybean_ir_sub_metrics)

# Example 2: Repeat the same process, only now plot background data as individual points. Note 
# this may be too slow now that all points are drawn in the background.

plotLitreApp(data = soybean_ir_sub_log, dataMetrics = soybean_ir_sub_metrics, option = "allPoints", 
  pointColor = "red")
}
}
