% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcanoApp.R
\name{plotVolcanoApp}
\alias{plotVolcanoApp}
\title{Plot interactive volcano plots}
\usage{
plotVolcanoApp(
  data = data,
  dataMetrics = dataMetrics,
  dataSE = NULL,
  option = c("hexagon", "allPoints"),
  pointColor = "orange"
)
}
\arguments{
\item{data}{DATA FRAME | Read counts}

\item{dataMetrics}{LIST | Differential expression metrics. This object must 
contain one column named "logFC" and one column named "PValue".}

\item{dataSE}{SUMMARIZEDEXPERIMENT | Summarized experiment format that
can be used in lieu of data; default NULL}

\item{option}{CHARACTER STRING ["hexagon" | "allPoints"] | The background of 
plot; default "hexagon"}

\item{pointColor}{CHARACTER STRING | Color of overlaid points on scatterplot 
matrix; default "orange"}
}
\value{
A Shiny application that shows a volcano plot and allows users to 
overlay genes depending on two values, usually a statistical value (such as
P-value) and a magnitude change value (such as log fold change). The user
can download a file that contains the gene IDs that pass these thresholds.
}
\description{
Plot interactive volcano plots.
}
\examples{
# The first pair of examples use data and dataMetrics objects as input.
# The last pair of examples create the same plots now using the 
# SummarizedExperiment (i.e. dataSE) object input.

# Example 1: Create interactive volcano plot of logged data using hexagon
# bins for the background.

data(soybean_cn_sub)
data(soybean_cn_sub_metrics)
app <- plotVolcanoApp(data = soybean_cn_sub,
    dataMetrics = soybean_cn_sub_metrics)
if (interactive()) {
    shiny::runApp(app)
}

# Example 2: Create interactive volcano plot of logged data using points for 
# the background.

app <- plotVolcanoApp(data = soybean_cn_sub,
    dataMetrics = soybean_cn_sub_metrics, option = "allPoints",
    pointColor = "magenta")
if (interactive()) {
    shiny::runApp(app)
}

# Below is the same pair of examples, only now using the
# SummarizedExperiment (i.e. dataSE) object as input.

# Example 1: Create interactive volcano plot of logged data using hexagon
# bins for the background.

\dontrun{
data(se_soybean_cn_sub)
app <- plotVolcanoApp(dataSE = se_soybean_cn_sub)
if (interactive()) {
    shiny::runApp(app)
}
}

# Example 2: Create interactive volcano plot of logged data using points for 
# the background.

\dontrun{
app <- plotVolcanoApp(dataSE = se_soybean_cn_sub, option = "allPoints",
    pointColor = "magenta")
if (interactive()) {
    shiny::runApp(app)
}
}

}
