% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcanoApp.R
\name{plotVolcanoApp}
\alias{plotVolcanoApp}
\title{Plot interactive volcano plots}
\usage{
plotVolcanoApp(data = data, dataMetrics = dataMetrics,
  option = "hexagon", pointColor = "orange")
}
\arguments{
\item{data}{DATA FRAME | Read counts}

\item{dataMetrics}{LIST | Differential expression metrics. This object must contain one column named "logFC" and one column named "PValue".}

\item{option}{CHARACTER STRING ["hexagon" | "allPoints"] | The background of plot; default "hexagon"}

\item{pointColor}{CHARACTER STRING | Color of overlaid points on scatterplot matrix; default "orange"}
}
\description{
Plot interactive volcano plots.
}
\examples{
\dontrun{
# Example 1: Create interactive volcano plot of logged data using hexagon bins for 
# the background.

data(soybean_cn_sub)
data(soybean_cn_sub_metrics)
plotVolcanoApp(data = soybean_cn_sub, dataMetrics = soybean_cn_sub_metrics)
}

\dontrun{
# Example 2: Create interactive volcano plot of logged data using points for the 
# background.

plotVolcanoApp(data = soybean_cn_sub, dataMetrics = soybean_cn_sub_metrics, 
  option = "allPoints", pointColor = "magenta")
}

}
